// WeddingService_AdminDashboard/src/app/modules/user/modules/super-admin/modules/vendor-dashboard/component/wallet-management/model/wallet-deduction.ts
export class WalletDeduction {
        leadId: number;
        leadIdNumber: string;
        leadMode: string;
        vendorId: number;
        customerId: number;
        categoryValue:  string;
        invoiceNumber: string;
        invoiceDate: Date;
        invoiceAmount: number;
        walletBalance: number;
        walletStatus: number;
        deductedAmount: number;
        deductedDate: Date;
        createdBy: number;
}